Update test to be runnable on Travis
The previous version of this test used a temporary file, which
TravisCI doesn't allow. This version uses a pipe to hold the
contents of stdout instead, and should therefore work on Travis.

That being said, we don't even need this test. This was just a
demo of capturing stdout in Go because I didn't know how to do it.
Once our code actually does something, we will delete this and
only look back if we need to do this at some point in the future.